requirements.txt
# SmartCityOS - Bibliotecas essenciais

# Banco de dados PostgreSQL
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2==2.9.11

# Visualização e gráficos
plotly==6.5.0
kaleido==1.2.0
Pillow==12.1.0

# PySide6 - Interface Desktop (Qt)
PySide6

# Dash GUI v1.1 - Interface Web
dash==2.18.2
dash-bootstrap-components==1.6.0

# Análise de dados e exportação
pandas==2.3.3
numpy==2.4.0
openpyxl==3.1.5
xlsxwriter==3.2.2
tabulate==0.9.0

# Configuração
python-dotenv==1.2.1

# Bibliotecas opcionais (descomente se desejar usar)
# tkinterweb==4.13.0          # Para renderização HTML inline
# pywebview==6.1             # Para janelas webview embutidas  
# cefpython3==66.1           # Para renderização com Chromium
